requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.4
dataclasses-json==0.5.14
diskcache==5.6.3
dnspython==2.4.2
fastapi==0.103.0
frozenlist==1.4.0
gptcache==0.1.42
guidance==0.0.64
h11==0.14.0
idna==3.4
langchain==0.0.274
langsmith==0.0.27
loguru==0.7.0
marshmallow==3.20.1
msal==1.24.1
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
nest-asyncio==1.5.7
ntplib==0.4.0
numexpr==2.8.5
numpy==1.25.2
openai==0.27.9
packaging==23.1
pandas==2.1.1
pinecone-client==2.2.2
platformdirs==3.11.0
protobuf==4.21.12
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pygtrie==2.5.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.4.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
yarl==1.9.2
